+++ packages/client/src/pages/CarbonHackCards/RegionRecommendationCard/RegionRecommendationCard.tsx
@@ -11,7 +11,6 @@ import TableHead from '@mui/material/TableHead'
 import TableRow from '@mui/material/TableRow'
 import Paper from '@mui/material/Paper'
 import { Typography, CardContent } from '@material-ui/core'
-import LoadingMessage from '../../../common/LoadingMessage'
 const RegionRecommendationCard: FunctionComponent<any> = ({
   data,
 }): ReactElement => {
@@ -19,6 +18,7 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
 
   const classes = useStyles()
   const footPrintData = data !== undefined ? data : []
+  console.log('Footprint: ', footPrintData)
 
   const findNearestRegions = (
     accountRegion: string,
@@ -31,6 +31,10 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
     } else {
       return ['us-east1', 'us-west1', 'us-east4', 'us-west2']
     }
+  }
+
+  const truncateDecimals = function (number) {
+    return Math.floor(number)
   }
 
   const getNearestRegionFromEnum = (accountRegion: string, REGIONS: any) => {
@@ -59,9 +63,10 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
   })
 
   const dataToSend = accountRegionMap
+  console.log(dataToSend, ' in emussins card')
 
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { result, error, loading } = useRegionRecommendationData(dataToSend)
+  const result = useRegionRecommendationData(dataToSend)
+  console.log(result, 'result of emission card')
 
   const bestLocationTable = (result) => {
     // eslint-disable-next-line prefer-const
@@ -69,57 +74,41 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
     for (const [key, value] of result) {
       if (value != undefined) {
         const expectedEmissionsForBestLocation =
-          (value.carbonIntensity * accountRegionMap.get(key)[3]) / 1000000
+          (truncateDecimals(value.carbonIntensity) *
+            truncateDecimals(accountRegionMap.get(key)[3])) /
+          1000000
         const actualEmissions = accountRegionMap.get(key)[2]
-        const percentReduction =
-          100 - (expectedEmissionsForBestLocation * 100) / actualEmissions
+        const percentReduction = truncateDecimals(
+          100 - (expectedEmissionsForBestLocation * 100) / actualEmissions,
+        )
         array.push(
-          <TableRow
-            key={actualEmissions}
-            sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
-          >
-            <TableCell align="center">{key}</TableCell>
-            <TableCell component="th" scope="row" align="center">
-              {accountRegionMap.get(key)[0]}
-            </TableCell>
-            <TableCell align="center">{actualEmissions.toFixed(5)}</TableCell>
-            <TableCell align="center">{value.location}</TableCell>
-            <TableCell align="center">
-              {accountRegionMap.get(key)[0] === value.location
-                ? actualEmissions.toFixed(5)
-                : expectedEmissionsForBestLocation.toFixed(5)}
-            </TableCell>
-            <TableCell align="center">
-              {accountRegionMap.get(key)[0] === value.location
-                ? 0
-                : Math.floor(percentReduction)}{' '}
-              %
-            </TableCell>
-          </TableRow>,
+          <>
+            <TableRow
+              sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
+            >
+              <TableCell align="center">{key}</TableCell>
+              <TableCell component="th" scope="row" align="center">
+                {accountRegionMap.get(key)[0]}
+              </TableCell>
+              <TableCell align="center">{actualEmissions}</TableCell>
+              <TableCell align="center">{value.location}</TableCell>
+              <TableCell align="center">
+                {accountRegionMap.get(key)[0] === value.location
+                  ? actualEmissions
+                  : expectedEmissionsForBestLocation}
+              </TableCell>
+              <TableCell align="center">
+                {accountRegionMap.get(key)[0] === value.location
+                  ? 0
+                  : percentReduction}{' '}
+                %
+              </TableCell>
+            </TableRow>
+          </>,
         )
       }
     }
     return array
-  }
-
-  const renderTable = () => {
-    return (
-      <TableContainer component={Paper}>
-        <Table sx={{ minWidth: 650 }} aria-label="simple table">
-          <TableHead>
-            <TableRow>
-              <TableCell align="center">Account</TableCell>
-              <TableCell align="center">Region</TableCell>
-              <TableCell align="center">Actual Emissions (Mg)</TableCell>
-              <TableCell align="center">Best Location</TableCell>
-              <TableCell align="center">Expected Emissions (Mg)</TableCell>
-              <TableCell align="center">Reduction in Emissions</TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>{bestLocationTable(result)}</TableBody>
-        </Table>
-      </TableContainer>
-    )
   }
 
   return (
@@ -139,13 +128,21 @@ const RegionRecommendationCard: FunctionComponent<any> = ({
           </Typography>
           <Typography className={classes.posOne}></Typography>
         </CardContent>
-        {loading ? (
-          <LoadingMessage
-            message={'Loading cloud data. This may take a while...'}
-          />
-        ) : (
-          renderTable()
-        )}
+        <TableContainer component={Paper}>
+          <Table sx={{ minWidth: 650 }} aria-label="simple table">
+            <TableHead>
+              <TableRow>
+                <TableCell align="center">Account</TableCell>
+                <TableCell align="center">Region</TableCell>
+                <TableCell align="center">Actual Emissions (Mg)</TableCell>
+                <TableCell align="center">Best Location</TableCell>
+                <TableCell align="center">Expected Emissions (Mg)</TableCell>
+                <TableCell align="center">Reduction in Emissions</TableCell>
+              </TableRow>
+            </TableHead>
+            <TableBody>{bestLocationTable(result)}</TableBody>
+          </Table>
+        </TableContainer>
       </React.Fragment>
     </DashboardCard>
   )
